fix: Avoid adding duplicate listeners for drag-n-drop on the web (#422)
**The problem**

I noticed that on Web DOM listeners are added again on each
reload/restart of the app and, during development, the number of events
increased rapidly.


https://github.com/user-attachments/assets/64d9217e-1057-4060-8f74-c496395462eb


https://github.com/user-attachments/assets/bbea0cfe-c022-4a92-9a59-3f2367ed14c3

**The solution**
We can't account for it in dart land very easily, because even the dart
context is thrown away and not even singletons survive. This is relevant
because the listeners need to be saved somewhere in order to be removed
later.

My solution is to attach the array of listeners as a DOM property, so
that in the next initialization we can retrieve and remove them all.

**Bonus**

Tests were added for this. Just run:
```sh
flutter test --platform chrome
```